requirements.txt
selenium~=4.26.1
undetected-chromedriver~=3.5.5
pillow~=11.0.0